Drop `EventListener.call_corr_xheader`
Remove the initial check for a special Xheader when associating sessions
into calls. Instead expect the user to assign `call_corr_var` to whatever
channel variable they desire (which can of course be an Xheader var).

The additional check for an Xheader var in `_handle_create` causes
a problem when a user wishes to associate sessions into calls using
some other channel var. For example with devices under test
which can't forward Xheaders a user may wish to use some other channel
variable like `Caller-Destination-Number`, however, since the originating
session would always contain the Xheader var (set by the originate command)
and the b-leg would not, the two sessions would never be associated into
a call. It is sensible and straightforward that only a single variable
should ever be used to index/identify calls by session.

By default the listener will still use `variable_call_uuid` since a listener
can be used purely to observe the FS slave's state and FS already uses
this variable to index calls bridged internally. When a `Client` is
assigned a listener, set the Xheader var ('sip_h_X-switchy_originating_session')
as the channel variable for call associations.
If Xheader forwarding is not supported by the DUT then the user
is expected to reassign this value.