Support preprending to the callbacks deque
Like it sounds; lets the user insert callbacks in front of all
others (i.e. so they're called first) for a given consumer id.
This turns out to be handy if we want some built in apps to
provide "core like" behaviour/state which can be consumed by
other apps. A prime example would be the `CallTimes` app which
records event time stamps for later metrics calculations.
Support this functionality via the `prepend` kwarg to
`Client.load_app`.

Additionally,
- strip out all event time stamp recording from default event
  handlers
- correct `EventListener.count_failed` to use the internal
  `hangup_causes` instance var instead of the size limited
  `failed_sessions` deque
- only pop calls in the job handler if a corresponding session
  can be looked up
- improve some handler logging